+++ packages/uikit/src/components/MenuItems/MenuItems.tsx
@@ -5,26 +5,7 @@ import { Flex } from "../Box";
 import DropdownMenu from "../DropdownMenu/DropdownMenu";
 import MenuItem from "../MenuItem/MenuItem";
 import { MenuItemsProps, MenuItemsType } from "./types";
-import { DropdownMenuItems } from "../DropdownMenu/types";
-import { Colors } from "../../theme";
-
-const findStatusColor = (items: DropdownMenuItems[] = []): keyof Colors | undefined => {
-  let statusColor: keyof Colors | undefined;
-
-  const traverseItems = (menuItems: DropdownMenuItems[]) => {
-    for (const item of menuItems) {
-      if (item.status?.color) {
-        statusColor = item.status.color;
-      }
-      if (item.items && item.items.length > 0) {
-        traverseItems(item.items);
-      }
-    }
-  };
-
-  traverseItems(items);
-  return statusColor;
-};
+import { findMenuItemsStatusColor } from "../../util/findMenuItemsStatusColor";
 
 const MenuItems: React.FC<React.PropsWithChildren<MenuItemsProps>> = ({
   items = [] as MenuItemsType[],
@@ -36,7 +17,7 @@ const MenuItems: React.FC<React.PropsWithChildren<MenuItemsProps>> = ({
   return (
     <Flex {...props}>
       {items.map(({ label, items: menuItems = [], href, icon, disabled, onClick }) => {
-        const statusColor = findStatusColor(menuItems);
+        const statusColor = findMenuItemsStatusColor(menuItems);
         const isActive = activeItem === href;
         const linkProps = isTouchDevice() && menuItems && menuItems.length > 0 ? {} : { href };
         const Icon = icon;
